packages/llm-tracekit/pyproject.toml
[project]
name = "llm-tracekit"
version = "1.7.0"
description = "Core package for LLM tracing with Coralogix integration."
readme = "README.md"
authors = [
    { name="Coralogix Ltd.", email="info@coralogix.com" },
]
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.9,<3.14"
dependencies = [
    "opentelemetry-api>=1.32.1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-instrumentation>=0.53b1",
    "opentelemetry-sdk>=1.32.1",
    "opentelemetry-semantic-conventions>=0.53b1",
    "pydantic>=2.0.0",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: OpenTelemetry :: Instrumentations",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
homepage = "https://coralogix.com"
repository = "https://github.com/coralogix/llm-tracekit.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
only-include = [
    "src",
]
